requirements.txt
faiss-cpu 
langchain 
sentence-transformers 
tiktoken 
openai
langgraph
streamlit
python-dotenv